+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Error translation moved to translation_utils.js
 * Export more functions to help constructing skins
+* trField to translate field names
 
 ### Fixed